Changes RecipientsMembers name. Uses MUI table
